+++ packages/f-checkout/CHANGELOG.md
@@ -5,10 +5,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v0.15.0
 ------------------------------
+<<<<<<< HEAD
 *December 2, 2020*
 
 ### Added
 - Webdriver IO tests for dropdown in `Selector` component.
+=======
+*December 3, 2020*
+
+### Added
+- Ability to change prop values in demo file for WebDriverIO component tests.
+- New `Demo.vue` for f-checkout controls.
+>>>>>>> master
 
 
 v0.14.0